Refractor AddItem to respect architecture
