+++ app/src/androidTest/java/com/github/brugapp/brug/QrCodeScannerActivityTest.kt
@@ -1,10 +1,7 @@
 package com.github.brugapp.brug
 
-import android.app.NotificationManager
-import android.content.Context
 import android.content.Intent
 import androidx.core.app.NotificationManagerCompat
-import androidx.core.content.ContextCompat.getSystemService
 import androidx.lifecycle.Lifecycle
 import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
