+++ app/src/androidTest/java/com/github/brugapp/brug/SettingsTest.kt
@@ -7,7 +7,6 @@ import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.brugapp.brug.ui.SettingsActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
