+++ app/src/main/java/com/github/brugapp/brug/fake/MockDatabase.kt
@@ -1,6 +1,5 @@
 package com.github.brugapp.brug.fake
 
-import com.github.brugapp.brug.model.Item
 import com.github.brugapp.brug.model.User
 
 class MockDatabase {
