+++ app/src/main/java/com/github/brugapp/brug/view_model/AddItemViewModel.kt
@@ -2,16 +2,28 @@ package com.github.brugapp.brug.view_model
 
 import android.widget.*
 import androidx.lifecycle.ViewModel
-import com.github.brugapp.brug.R
 import com.github.brugapp.brug.fake.MockDatabase.Companion.currentUser
 import com.github.brugapp.brug.fake.MockDatabase.Companion.itemId
 import com.github.brugapp.brug.model.Item
+import com.github.brugapp.brug.model.ItemType
 
 class AddItemViewModel : ViewModel() {
 
-    fun addItem(itemNameView : EditText, description : EditText){
-        val newItem = Item(itemNameView.text.toString(), description.text.toString(), itemId)
+    fun addItem(itemNameView: EditText, description: EditText, itemType: Spinner){
+        val newItem = Item(itemNameView.text.toString(), description.text.toString(),itemId)
+            .setType(stringToItemType(itemType.selectedItem.toString()))
         currentUser.addItem(newItem)
+    }
+
+    private fun stringToItemType(str:String): ItemType {
+        val type = when (str){
+            "Wallet" -> ItemType.Wallet
+            "Keys" -> ItemType.Keys
+            "Car keys" -> ItemType.CarKeys
+            "Phone" -> ItemType.Phone
+            else -> ItemType.Other
+        }
+        return type
     }
 
     fun verifyForm(nameHelperText : TextView, itemNameView : EditText) : Boolean{
@@ -35,7 +47,7 @@ class AddItemViewModel : ViewModel() {
         return true
     }
 
-     private fun validName(itemNameView : EditText): String {
+    private fun validName(itemNameView : EditText): String {
         val nameText = itemNameView.text.toString()
 
         return if(nameText.isEmpty()) "Name must contain at least 1 character" else ""
